Automatically Committed for MakeGamesWithUs

+++ GameOfLife.spritebuilder/Grid.m
@@ -136,7 +136,7 @@ static const int GRID_COLUMNS = 10;
     for (int i = 0; i < [_gridArray count]; i++) {
         for (int j = 0 ; j < [_gridArray[i]count]; j++) {
             Creature * currentCreature = _gridArray[i][j];
-            int currentAlive = currentCreature.livingNeigbors;
+            int currentAlive = (int)currentCreature.livingNeigbors;
             if (currentAlive == 3) {
                 currentCreature.isAlive = YES;
             } else if ((currentAlive <= 1)||(currentAlive >= 4)) {
